+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - set up automatic dependency injection on commands
 - update `ConfigureAppHook` to load configs from every bundle in the hierarchy
 - require `sqlalchemy-unchained>=0.8`
+- add `flask unchained config` command for listing the current config (optionally filtered by bundle)
+- add `flask unchained extensions` command for listing extensions discovered by the app
+- add `flask unchaiend services` command for listing services discovered by the app
 
 ## v0.7.9 (2019/05/19)